+++ NEWS.md
@@ -1,3 +1,8 @@
+# RprobitB 1.1.2
+
+* `logLik()` method now returns a `logLik` object that can be passed to `AIC()` and `BIC()` methods from {stats}.
+* Small bug fixes for the ordered probit model case.
+
 # RprobitB 1.1.1
 
 * Small documentation fixes.